refactor: tag status as extract style
